requirements.txt
PyYAML>=6.0.1
PyGObject>=3.42,<3.43
pycairo>=1.20.0
matplotlib>=3.8,<3.10
